Change ThreadMethods type from long[] to int[] in Compiler
Updated the ThreadMethods array in the Compiler class to use int[]
instead of long[]. This reduces memory usage and is sufficient if
only integer values are required for tracking thread methods.

+++ Source/Mosa.Compiler.Framework/Compiler.cs
@@ -30,7 +30,7 @@ public sealed class Compiler
 	private int ActiveThreadCount;
 	private long[] ThreadCPUTicks;
 	private long[] ThreadWallTicks;
-	private long[] ThreadMethods;
+	private int[] ThreadMethods;
 
 	#endregion Data Members
 
@@ -250,7 +250,7 @@ public sealed class Compiler
 		MethodStagePipelines = new Pipeline<BaseMethodCompilerStage>[MaxThreads];
 		ThreadCPUTicks = new long[MaxThreads];
 		ThreadWallTicks = new long[MaxThreads];
-		ThreadMethods = new long[MaxThreads];
+		ThreadMethods = new int[MaxThreads];
 
 		MethodScheduler = new MethodScheduler(this);
 		MethodScanner = new MethodScanner(this);
